requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.9.1
beautifulsoup4==4.12.3
cached-property==1.5.2
certifi==2024.2.2
cffi==1.16.0
chardet==4.0.0
cryptography==3.4.8
DateTime==5.5
decorator==4.4.2
docopt==0.6.2
docutils==0.17
ebaysdk==2.1.5
exceptiongroup==1.2.1
freezegun==1.1.0
frozenlist==1.4.1
geoip2==2.9.0
gevent==21.8.0
greenlet==1.1.2
h11==0.14.0
html2text==2024.2.25
idna==2.10
isodate==0.6.1
Jinja2==3.0.3
libsass==0.20.1
lxml==4.8.0
MarkupSafe==2.0.1
maxminddb==2.6.1
multidict==6.0.5
num2words==0.5.10
numpy==2.1.3
ofxparse==0.21
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.3
passlib==1.7.4
pgeocode==0.3.0
Pillow==9.0.1
platformdirs==4.2.1
polib==1.1.1
psutil==5.9.0
psycopg2==2.9.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydot==1.4.2
pyOpenSSL==21.0.0
pyparsing==3.1.2
PyPDF2==1.26.0
pyserial==3.5
PySocks==1.7.1
pytesseract==0.3.10
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-ldap==3.4.0
python-stdnum==1.17
pytz==2024.1
pyusb==1.2.1
qrcode==7.3.1
razorpay==1.4.2
reportlab==3.6.8
requests==2.25.1
requests-file==2.0.0
requests-toolbelt==1.0.0
rjsmin==1.1.0
selenium==4.22.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.5
vobject==0.9.6.1
websocket-client==1.8.0
Werkzeug==2.0.2
wsproto==1.2.0
xlrd==1.2.0
XlsxWriter==3.0.2
xlwt==1.3.0
yarl==1.9.4
zeep==4.1.0
zope.event==5.0
zope.interface==6.4.post2
